trainer controller fixes, passrate fixes, mapper fixes

+++ appSchool/src/main/java/com/github/odyn666/appSchool/dto/TrainerEntityDto.java
@@ -28,7 +28,7 @@ public final class TrainerEntityDto implements Serializable {
     private final String identifier;
     private final String phoneNumber;
     private final String email;
-    private final Integer studentsPassRate;
+    private final Float studentsPassRate;
     private final List<TrainerOpinions> trainerOpinions;
